增加了Array.prototype.InsertAfter.InsertAfter
便于寻找和插入特定位置

+++ service.js
@@ -1,3 +1,18 @@
+
+var Errhandler = function(err){
+    console.error(err);
+    if (err.data) {
+        if (err.data.msg) {
+            alert(err.data.msg);
+        }else {
+            alert(err.data);
+        }
+    }else {
+        alert(err);
+    }
+}
+
+
 app.factory('ImageServ', function() {
     var instance = {
 
@@ -91,7 +106,7 @@ app.provider('Dialogs', function() {
         return {
             Confirm: function(content, title) {
                 var win = $modal.open({
-                    templateUrl: '/js/tpls/dialog-confirm.html',
+                    templateUrl: '/zjs/tpls/dialog-confirm.html',
                     controller: function($scope, $modalInstance, params) {
                         $scope.title = params.title || '确认';
                         $scope.content = params.content;
@@ -114,7 +129,7 @@ app.provider('Dialogs', function() {
 
             Form: function(scope, params) {
                 var win = $modal.open({
-                    templateUrl: '/js/tpls/dialog-form.html',
+                    templateUrl: '/zjs/tpls/dialog-form.html',
                     controller: function($scope, $modalInstance, params) {
                         angular.extend($scope, params);
 
@@ -223,14 +238,14 @@ app.provider('ControllerHelper', function(){
 
 
                 $scope.Delete = function(data) {
-                    // return Dialogs.Confirm("是否确认删除此数据?", "删除确认")
-                    // .then(function(isYes){
-                    //     if (isYes) {
+                    return Dialogs.Confirm("是否确认删除此数据?", "删除确认")
+                    .then(function(isYes){
+                        if (isYes) {
                             return rs.DeleteById(data._id).then(function(){
                                 $state.go(listState);
                             }, Errhandler);
-                    //     }
-                    // })
+                        }
+                    })
                 }
 
                 $scope.Save = function(data) {
@@ -248,12 +263,15 @@ app.provider('ControllerHelper', function(){
                     //    rstCondition[restrict] = $state.params[restrict];
                     //});
                     return rs.Save(data).then(function(su){
+                        // 更新当前数据
+                         InitSetCurItem();
+
                         //显示提示信息
                         if (!$rootScope.alerts)
                             $rootScope.alerts = [];
                         var alert = {msg:'数据保存完毕。', type:'success'};
                         $rootScope.alerts.push(alert);
-                        // 几秒钟后，自动小时。
+                        // 几秒钟后，自动消失。
                         $timeout(function(){
                             $rootScope.alerts.Delete(alert);
                         }, 3000);
@@ -271,7 +289,12 @@ app.provider('ControllerHelper', function(){
                 }
 
                 $scope.DialogEdit = function(data, templateUrl) {
-                    var templateUrl = templateUrl || $state.get(editState).views[""].templateUrl;
+                    if (!templateUrl)  {
+                        var state = $state.get(editState);
+                        if (!state)
+                            throw new Error('Can not get state: '  + editState);
+                        templateUrl = $state.get(editState).views[""].templateUrl;
+                    }
 
                     $scope[formModelName] = angular.copy(data);
                     Dialogs.Form($scope, {
@@ -296,7 +319,8 @@ app.provider('ControllerHelper', function(){
 
                 return $q.when()
                     .then(InitLoadList)
-                    .then(InitSetCurItem);
+                    .then(InitSetCurItem)
+                    .catch(Errhandler);
             }
         }
     }
@@ -344,6 +368,12 @@ app.factory('DictServ', function($rootScope) {
             {k: 'Supplier', v: '供应商'},
             {k: 'Client', v: '客户'},
             {k: 'Other', v: '其他'},
+        ]);
+        CreateDict('LogiType', [
+            {k: 'agent', v: '货运代理'},
+            {k: 'logistics', v: '跨国货运公司'},
+            {k: 'local', v: '同城快递'},
+            {k: 'national', v: '国内快递'},
         ]);
     };
 
@@ -476,7 +506,6 @@ app.filter('removeExistsArrayBy_id', function() {
 });
 
 if (typeof CartRS !== 'undefined') {
-
     app.provider('CartServ', function(){
         this.$get = function($rootScope, $state, $cookieStore , CartRS, $timeout, $q, Cache, Dialogs) {
             var CreateCart = function() {
@@ -555,6 +584,7 @@ if (typeof CartRS !== 'undefined') {
 
                 Items: function() {
                     return cartItems;
+
                 },
 
                 TotalPrice: function() {
@@ -566,7 +596,7 @@ if (typeof CartRS !== 'undefined') {
                 },
 
                 IncQuantity: function(item, n) {
-                    var res = item.quantity + n
+                    var res = item.quantity + n;
                     if (res <= 1)
                         item.quantity = 1;
                     else
@@ -574,11 +604,12 @@ if (typeof CartRS !== 'undefined') {
                     SaveCart();
                 },
             }
-        }
+        };
     });
-
-    app.run(function($rootScope, CartServ){
-        $rootScope.CartServ = CartServ;
-    });
-
 }
+
+//TODO 关于CartServ这种动态加载的Serv如何去实现，还没有解决！！
+//app.run(function($rootScope, CartServ){
+    //$rootScope.CartServ = CartServ;
+app.run(function($rootScope){
+});
